# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kraken Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface KeysStatus {
  configured: boolean;
  paperMode: boolean;
}

export interface ApiKeysInput {
  apiKey: string;
  apiSecret: string;
}

export type SettingsMode = typeof SettingsMode[keyof typeof SettingsMode];


export const SettingsMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface Settings {
  /** Percentage of USD balance to allocate (0-100) */
  allocation: number;
  mode: SettingsMode;
  /** Minimum profit % to take (default 5) */
  profitTarget: number;
  /** Trailing stop loss % (default 2) */
  trailingStop: number;
  /** Max trades at once (default 2) */
  maxConcurrentTrades: number;
  /** Minimum votes needed to trigger a trade (default 4) */
  voteThreshold: number;
}

export type SettingsInputMode = typeof SettingsInputMode[keyof typeof SettingsInputMode];


export const SettingsInputMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface SettingsInput {
  allocation?: number;
  mode?: SettingsInputMode;
  profitTarget?: number;
  trailingStop?: number;
  maxConcurrentTrades?: number;
  voteThreshold?: number;
}

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  open: 'open',
  closed: 'closed',
  stopped: 'stopped',
} as const;

/**
 * Why the trade was closed
 */
export type TradeCloseReason = typeof TradeCloseReason[keyof typeof TradeCloseReason] | null;


export const TradeCloseReason = {
  profit: 'profit',
  stop: 'stop',
  sell_signal: 'sell_signal',
  manual: 'manual',
  swapped: 'swapped',
} as const;

export interface Trade {
  id: string;
  symbol: string;
  name: string;
  entryPrice: number;
  currentPrice: number;
  quantity: number;
  investedUsd: number;
  profitPercent: number;
  profitUsd: number;
  status: TradeStatus;
  strategy: string;
  winningStrategies: string[];
  openedAt: string;
  closedAt?: string | null;
  paperMode: boolean;
  highestPrice: number;
  trailingActive: boolean;
  /** Voting confidence score (0-1) at the time this trade was opened */
  entryConfidence: number;
  /** Why the trade was closed */
  closeReason?: TradeCloseReason;
}

export interface TradeList {
  trades: Trade[];
  total: number;
  openCount: number;
  closedCount: number;
}

export type StrategyCurrentSignal = typeof StrategyCurrentSignal[keyof typeof StrategyCurrentSignal];


export const StrategyCurrentSignal = {
  buy: 'buy',
  sell: 'sell',
  hold: 'hold',
  inactive: 'inactive',
} as const;

export interface Strategy {
  id: string;
  name: string;
  description: string;
  /** Current learned weight (0-1) */
  weight: number;
  winRate: number;
  totalSignals: number;
  successfulSignals: number;
  profitContribution: number;
  currentSignal: StrategyCurrentSignal;
}

export interface StrategiesResponse {
  strategies: Strategy[];
  learningCycles: number;
  lastUpdated: string;
}

export type StrategyVoteVote = typeof StrategyVoteVote[keyof typeof StrategyVoteVote];


export const StrategyVoteVote = {
  buy: 'buy',
  sell: 'sell',
  hold: 'hold',
} as const;

export interface StrategyVote {
  strategyId: string;
  strategyName: string;
  vote: StrategyVoteVote;
  confidence: number;
  weight: number;
}

export type VoteResultDecision = typeof VoteResultDecision[keyof typeof VoteResultDecision];


export const VoteResultDecision = {
  buy: 'buy',
  sell: 'sell',
  hold: 'hold',
} as const;

export interface VoteResult {
  symbol: string;
  name: string;
  price: number;
  totalWeightedVotes: number;
  buyScore: number;
  sellScore: number;
  holdScore: number;
  decision: VoteResultDecision;
  confidence: number;
  votes: StrategyVote[];
}

export interface VotesResponse {
  results: VoteResult[];
  timestamp: string;
}

export type MarketTickerCategory = typeof MarketTickerCategory[keyof typeof MarketTickerCategory];


export const MarketTickerCategory = {
  crypto: 'crypto',
  meme: 'meme',
} as const;

export interface MarketTicker {
  symbol: string;
  krakenPair: string;
  name: string;
  price: number;
  change24h: number;
  volume24h: number;
  high24h: number;
  low24h: number;
  category: MarketTickerCategory;
}

export interface TickerResponse {
  tickers: MarketTicker[];
  lastUpdated: string;
}

export type BotStatusMode = typeof BotStatusMode[keyof typeof BotStatusMode];


export const BotStatusMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface BotStatus {
  running: boolean;
  mode: BotStatusMode;
  activeTradeCount: number;
  activeTrades: Trade[];
  balance: number;
  allocatedBalance: number;
  lastScanAt?: string | null;
  scanIntervalSeconds: number;
}

export interface Portfolio {
  totalBalance: number;
  availableBalance: number;
  allocatedBalance: number;
  allocatedInTrades: number;
  totalPnl: number;
  totalPnlPercent: number;
  winRate: number;
  totalTrades: number;
  paperMode: boolean;
}

export type CloseTrade404 = {
  error?: string;
};

export type GetTradesParams = {
limit?: number;
offset?: number;
mode?: GetTradesMode;
};

export type GetTradesMode = typeof GetTradesMode[keyof typeof GetTradesMode];


export const GetTradesMode = {
  paper: 'paper',
  live: 'live',
  all: 'all',
} as const;