# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kraken Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades/{id}/close:
    post:
      operationId: closeTrade
      tags: [trades]
      summary: Manually close an open trade at current market price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade closed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Trade not found or already closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Get trade history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: mode
          in: query
          schema:
            type: string
            enum: [paper, live, all]
            default: all
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        entryPrice:
          type: number
        currentPrice:
          type: number
        quantity:
          type: number
        investedUsd:
          type: number
        profitPercent:
          type: number
        profitUsd:
          type: number
        status:
          type: string
          enum: [open, closed, stopped]
        strategy:
          type: string
        winningStrategies:
          type: array
          items:
            type: string
        openedAt:
          type: string
        closedAt:
          type: string
          nullable: true
        paperMode:
          type: boolean
        highestPrice:
          type: number
        trailingActive:
          type: boolean
        entryConfidence:
          type: number
          description: Voting confidence score (0-1) at the time this trade was opened
        closeReason:
          type: string
          nullable: true
          enum: [profit, stop, sell_signal, manual, swapped]
          description: Why the trade was closed
      required: [id, symbol, name, entryPrice, currentPrice, quantity, investedUsd, profitPercent, profitUsd, status, strategy, winningStrategies, openedAt, paperMode, highestPrice, trailingActive, entryConfidence]
    TradeList:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
        openCount:
          type: integer
        closedCount:
          type: integer
      required: [trades, total, openCount, closedCount]
...