# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kraken Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeCloseReason } from './tradeCloseReason';
import type { TradeStatus } from './tradeStatus';

export interface Trade {
  id: string;
  symbol: string;
  name: string;
  entryPrice: number;
  currentPrice: number;
  quantity: number;
  investedUsd: number;
  profitPercent: number;
  profitUsd: number;
  status: TradeStatus;
  strategy: string;
  winningStrategies: string[];
  openedAt: string;
  closedAt?: string | null;
  paperMode: boolean;
  highestPrice: number;
  trailingActive: boolean;
  /** Voting confidence score (0-1) at the time this trade was opened */
  entryConfidence: number;
  /** Why the trade was closed */
  closeReason?: TradeCloseReason;
}